Extract Prompt Studio legend helpers
